requirements.txt
streamlit==1.32.2
pandas==2.2.1
joblib==1.3.2
fpdf==1.7.2
openai==1.79.0
langchain==0.3.9
faiss-cpu==1.9.0.post1
scikit-learn==1.3.2
langchain-openai==0.3.17
langchain-community==0.3.9
langchain-core==0.3.60
python-dotenv

